backend/requirements.txt
annotated-types
anyio
azure-ai-inference
azure-core
blinker
certifi
charset-normalizer
click
distro
dnspython
Flask
flask-cors
h11
httpcore
httpx
idna
isodate
itsdangerous
Jinja2
MarkupSafe
pymongo
python-dotenv
Werkzeug
bcrypt
PyJWT
